add topic delete function
